+++ systems/__init__.py
@@ -1,2 +1,0 @@
-from .Mod00 import Mod00
-from .Mod01 import Mod01