+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2022/07/04 - 00 - Separated Systems
+> Toolbox version 0.8.5
+* Added data and scripts for `v2.0` of the manuscript.
+* Updated `README`.
+
 ## 2021/10/28 - 00 - Version Upgrade
 > Toolbox version 0.8.0
 * Added adiabatic elimination notebook for `mod_00`.